pyproject.toml
[project]
name = "test-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "PeterGQ",email = "pandeve3@gmail.com"}
]
readme = "README.md"
requires-python = "^3.9"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "flask-cors (>=6.0.1,<7.0.0)",
    "flask-jwt-extended (>=4.7.1,<5.0.0)",
    "sentry-sdk[flask] (>=2.35.2,<3.0.0)",
    "prometheus-flask-exporter (>=0.23.2,<0.24.0)",
    "python-json-logger (>=3.3.0,<4.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "stripe (>=12.5.1,<13.0.0)"
]

[tool.poetry]
packages = [{include = "test_backend", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
